% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Physical measurements of abalone.}
\format{A data frame with 4777 rows and 9 variables.
\describe{
  \item{Sex}{Sex: M, F, and I (infant)}
  \item{Length}{Longest shell measurement, in mm}
  \item{Diameter}{perpendicular to length, in mm}
  \item{Height}{with meat in shell, in mm}
  \item{Whole_weight}{whole abalone, in grams}
  \item{Shucked_weight}{weight of meat, in grams}
  \item{Viscera_weight}{gut weight (after bleeding), in grams}
  \item{Shell_weight}{after being dried, in grams}
  \item{Class}{the age in class}
}}
\usage{
abalone
}
\description{
A dataset containing several physical measurements of abalone and their ages
(coded in two classes).
Originally from \url{http://archive.ics.uci.edu/ml/datasets/Abalone}
}
\keyword{datasets}
