% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{acc}
\alias{acc}
\alias{mtr_accuracy}
\alias{mtr_ccr}
\title{Accuracy}
\usage{
mtr_accuracy(actual, predicted, cutoff = 0.5)

mtr_ccr(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}
}
\value{
A numeric scalar output
}
\description{
Accuracy
}
\seealso{
\code{\link{mtr_balanced_accuracy}}
}
\author{
An Chu
}
