% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{accuracy}
\alias{accuracy}
\title{Compute the accuracy}
\usage{
accuracy(actual, predicted)
}
\arguments{
\item{actual}{The ground truth vector.}

\item{predicted}{The predicted vector.}
}
\description{
This function computes the accuracy score between two vectors
}
\examples{
actual <- c('a', 'a', 'c', 'b', 'c')
predicted <- c('a', 'b', 'c', 'b', 'a')
accuracy(actual, predicted)
}
