% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc.r
\name{accuracy_ratio}
\alias{accuracy_ratio}
\alias{mtr_ar}
\alias{mtr_accuracy_ratio}
\alias{mtr_gini_coef}
\title{Accuracy Ratio}
\usage{
mtr_ar(actual, predicted)

mtr_accuracy_ratio(actual, predicted)

mtr_gini_coef(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}
}
\value{
A numeric scalar output
}
\description{
\code{mtr_ar}
}
\examples{

set.seed(101)
pred <- runif(1000)
act <- round(pred)
pred[sample(1000, 500)] <- runif(500)   # noise
mtr_accuracy_ratio(act, pred)


}
\author{
An Chu
}
