\name{auc}
\alias{auc}
\title{Compute the area under the ROC (AUC)}
\usage{
  auc(actual, predicted)
}
\arguments{
  \item{actual}{binary vector}

  \item{predicted}{real-valued vector that defines the
  ranking}
}
\description{
  This function computes the area under the
  receiver-operator characteristic (AUC)
}

