% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{ce}
\alias{ce}
\title{Compute the classification error}
\usage{
ce(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the classification error
between two vectors
}

