% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{classification_params}
\alias{classification_params}
\title{Classification Metrics Parameters}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}
}
\description{
Documentation for shared parameters of functions that computes
classification metrics.
}
