% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_lift}
\alias{compute_lift}
\title{Compute the lift}
\usage{
compute_lift(data, obs = "obs", pos_class = "Yes",
  quantile = "quantile", nb_quantiles = 20)
}
\arguments{
\item{data}{A dataframe with the observed class, the probabilities for the
positive class and the quantiles.}

\item{obs}{A string (default = "obs"). The column's name of the observed
class.}

\item{pos_class}{A string (default = "Yes"). How the positive class is coded.}

\item{quantile}{A string (default = "quantile"). The column's name of the
quantiles.}

\item{nb_quantiles}{An integer (default = 20). The number of quantiles
computed.}
}
\value{
A dataframe with the lift and the cumulative lift computed for each
  quantile.
}
\description{
Compute the lift and cumulative lift.
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  prob = c(runif(n = 20, min = 0.3, max = 0.8),
           runif(n = 20, min = 0.1, max = 0.6))
)

quantiles <- qcut(data, prob = "prob")
lift <- compute_lift(quantiles)
}
\seealso{
\code{\link{compute_lift}}, \code{\link{draw_lift}}
}
