% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute several metrics for classification tasks.}
\usage{
compute_metrics(data, obs = "obs", pred = "pred", pos_class = "Yes",
  neg_class = "No")
}
\arguments{
\item{data}{A dataframe with the observed and the predicted class.}

\item{obs}{A string (default = "obs"). The column's name of the observed
class.}

\item{pred}{A string (default = "pred"). The column's name of the predicted
class.}

\item{pos_class}{A string (default = "Yes"). How the positive class is coded.}

\item{neg_class}{A string (default = "No"). How the negative class is coded.}
}
\value{
A dataframe with the computed metrics.
}
\description{
Compute the following metrics: "Accuracy", "Precision", "Recall",
"Specificity", "NPV" and "F1".
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  Yes = c(runif(n = 20, min = 0.3, max = 0.8),
          runif(n = 20, min = 0.1, max = 0.6))
)

predict_class <- compute_predict_class(data, pos_class_col = "Yes")

compute_metrics(predict_class)
}
\seealso{
\code{\link{compute_predict_class}},
  \code{\link{compute_optimal_threshold}}, \code{\link{objective_threshold}}
}
