% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_optimal_threshold}
\alias{compute_optimal_threshold}
\title{Compute the optimal threshold for the accuracy or the F1-score}
\usage{
compute_optimal_threshold(data, metric = c("Accuracy", "F1"))
}
\arguments{
\item{data}{A dataframe with the metrics computed for different thresholds.}

\item{metric}{A string. The metric to maximize. One of "Accuracy" or "F1".}
}
\value{
A dataframe with the optimal threshold(s).
}
\description{
Compute the optimal threshold for the accuracy or the F1-score.
}
\examples{
# add an example here
}
\seealso{
\code{\link{compute_predict_class}}, \code{\link{compute_metrics}},
  \code{\link{objective_threshold}}
}
