% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{confusion_matrix}
\alias{confusion_matrix}
\alias{confusion_list}
\alias{mtr_confusion_matrix}
\title{Confusion Matrix}
\usage{
confusion_list(actual, predicted, cutoff = 0.5)

mtr_confusion_matrix(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}
}
\value{
A named list or a named two dimensions matrix
}
\description{
Confusion matrix - a special kind of contingency table - is a specific table
layout that allows the visualization of performance of a classification
model (or classifier). It composes of four different combinations of
predicted and actual values.

\code{confusion_list} computes and returns those combinations as a
\code{named list} while \code{mtr_confusion_matrix} returns a \code{named
matrix}.

Here's elements of confusion matrix:

\itemize{
\item True Positive (TP)
\item False Positive (FP)
\item True Negative (TN)
\item False Negative (FN)
}
}
\examples{

act <- c(1, 0, 0, 1, 1)
pred <- c(0.9, 0.3, 0.6, 0.5, 0.2)

## output as a R's list
## metrics:::confusion_list(act, pred) # default value of cutoff = 0.5
## metrics:::confusion_list(act, pred, cutoff = 0.7)

## output as a R's matrix
mtr_confusion_matrix(act, pred)
mtr_confusion_matrix(act, pred, cutoff = 0.7)


}
\author{
An Chu
}
