% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbif_utils.R
\name{convert_country_name_to_code}
\alias{convert_country_name_to_code}
\title{Convert country name to ISO 3166-1 alpha-2 country code}
\usage{
convert_country_name_to_code(country_name)
}
\arguments{
\item{country_name}{Character string. The full name of the country (e.g., "United States").}
}
\value{
A character string representing the ISO 3166-1 alpha-2 country code (e.g., "US").
}
\description{
Converts the country name (e.g., "United States") to its ISO 3166-1 alpha-2 country code (e.g., "US").
}
