% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_inputs.R
\name{convert_ecological_impact}
\alias{convert_ecological_impact}
\title{Convert user input for ecological impact to a numeric value}
\usage{
convert_ecological_impact(input)
}
\arguments{
\item{input}{A character string representing the selected ecological impact option.}
}
\value{
A numeric value corresponding to the selected ecological impact option:
1 for "Some population fitness decline",
2 for "Some population extermination",
3 for "Community loss",
4 for "test_option_4",
5 for "test_option_5".
}
\description{
This function converts the input selected by the user regarding the impact of a species on ecology
to a numeric value. The input is expected to be one of the predefined options.
}
