% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_inputs.R
\name{convert_ecosystem_impact}
\alias{convert_ecosystem_impact}
\title{Convert user input for ecosystem impact to a numeric value}
\usage{
convert_ecosystem_impact(input)
}
\arguments{
\item{input}{A character string representing the selected ecosystem impact option.}
}
\value{
A numeric value corresponding to the selected ecosystem impact option:
1 for "Replaces functions",
2 for "Hinders functions",
3 for "LOSS of functions",
4 for "test_option_4",
5 for "test_option_5".
}
\description{
This function converts the input selected by the user regarding the impact of a species on ecosystem
functioning to a numeric value. The input is expected to be one of the predefined options.
}
