% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_inputs.R
\name{convert_health_impact}
\alias{convert_health_impact}
\title{Convert user input for health impact to a numeric value}
\usage{
convert_health_impact(input)

convert_health_impact(input)
}
\arguments{
\item{input}{A character string representing the selected health impact option.}
}
\value{
A numeric value corresponding to the selected health impact option:
1 for "Nuisance",
2 for "Can't hurt/kill individuals",
3 for "Can threaten populations",
4 for "test_option_4",
5 for "test_option_5".

A numeric value corresponding to the selected health impact option:
1 for "Nuisance",
2 for "Can't hurt/kill individuals",
3 for "Can threaten populations",
4 for "test_option_4",
5 for "test_option_5".
}
\description{
This function converts the input selected by the user regarding the impact of a species on human health
to a numeric value. The input is expected to be one of the predefined options.

This function converts the input selected by the user regarding the impact of a species on human health
to a numeric value. The input is expected to be one of the predefined options.
}
