% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_inputs.R
\name{convert_reversibility_impact}
\alias{convert_reversibility_impact}
\title{Convert user input for reversibility impact to a numeric value}
\usage{
convert_reversibility_impact(input)
}
\arguments{
\item{input}{A character string representing the selected reversibility impact option.}
}
\value{
A numeric value corresponding to the selected reversibility impact option:
1 for "Not noticeable",
2 for "Reversible (some species decline)",
3 for "Irreversible (some species die out)",
4 for "test_option_4",
5 for "test_option_5".
}
\description{
This function converts the input selected by the user regarding the reversibility impact of a species
to a numeric value. The input is expected to be one of the predefined options.
}
