% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_inputs.R
\name{convert_social_perception}
\alias{convert_social_perception}
\title{Convert social perception input to a numeric value}
\usage{
convert_social_perception(input)
}
\arguments{
\item{input}{A character string representing the social perception.}
}
\value{
A numeric value corresponding to the selected option:
1 for "People Don't Like It",
2 for "People Don't Care",
3 for "People Like It",
4 for "test_option_4",
5 for "test_option_5".
}
\description{
This function takes a user input for social perception and converts it to a corresponding numeric value.
The function expects a specific set of string inputs and returns an associated numeric value.
If the input is invalid, it throws an error.
}
