% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{draw_lift}
\alias{draw_lift}
\title{Plot the lift curves}
\usage{
draw_lift(data, type = c("gain_chart", "lift_curve",
  "cumulative_lift_curve"), quantile = "quantile", nb_quantiles = 20,
  lift = "lift", cumulative_lift = "cumulative_lift",
  cumulative_lift_per = "cumulative_lift_per",
  cumulative_lift_per_ideal = "cumulative_lift_per_ideal")
}
\arguments{
\item{data}{A dataframe or a list of dataframes with the lift and the
cumulative lift per quantile.}

\item{type}{A string (default = "gain_chart"). The type of plot. One of:
"gain_chart", "lift_curve" or "cumulative_lift_curve".}

\item{quantile}{A string (default = "quantile"). The column's name of the
quantiles.}

\item{nb_quantiles}{An integer (default = 20). The number of quantiles
computed.}

\item{lift}{A string (default = "lift"). The column's name of the lift}

\item{cumulative_lift}{A string (default = "cumulative_lift"). The column's
name of the cumulative lift.}

\item{cumulative_lift_per}{A string (default = "cumulative_lift_per"). The
column's name of the cumulative lift in percentage.}

\item{cumulative_lift_per_ideal}{A string (default =
"cumulative_lift_per_ideal"). The column's name of the cumulative lift in
percentage for the perfect model.}
}
\value{
A ggplot2 object.
}
\description{
Plot the lift curves.
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  prob = c(runif(n = 20, min = 0.3, max = 0.8),
           runif(n = 20, min = 0.1, max = 0.6))
)

quantiles <- qcut(data, prob = "prob")
lift <- compute_lift(quantiles)

draw_lift(lift)
}
\seealso{
\code{\link{compute_lift}}, \code{\link{draw_lift}}
}
