% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{draw_roc}
\alias{draw_roc}
\title{Plot the ROC curve}
\usage{
draw_roc(data, recall = "Recall", specificity = "Specificity")
}
\arguments{
\item{data}{A dataframe or a list of dataframes with the recall and the
specificity.}

\item{recall}{A string (default = "Recall"). The column's name of the recall.}

\item{specificity}{A string (default = "Specificity"). The column's name of
the specificity.}
}
\value{
A ggplot2 object.
}
\description{
Plot the ROC curve.
}
\examples{
# add an example here
}
\seealso{
\code{\link{compute_predict_class}}, \code{\link{compute_metrics}}
}
