% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{explained_variation}
\alias{explained_variation}
\title{Explained Variation}
\usage{
explained_variation(actual, predicted)
}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector
is a prediction for the corresponding element in \code{actual}.}
}
\description{
\code{explained_variation} computes the percentage of variation in one numeric vector
explained by another, also known as the coefficient of determination.
}
\details{
\code{explained_variation} subtracts the relative squared error, \code{rse(actual, predicted)},
from 1, meaning it can return negative values if the predictions are on average further from
the actual values than predictions from a naive model that predicts the mean for every data
point.
}
\examples{
actual <- c(1.1, 1.9, 3.0, 4.4, 5.0, 5.6)
predicted <- c(0.9, 1.8, 2.5, 4.5, 5.0, 6.2)
explained_variation(actual, predicted)
}
\seealso{
\code{\link{rse}}
}
