% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{f1}
\alias{f1}
\title{Compute the f1 score
This function computes the f1 score between
two vectors.}
\usage{
f1(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
Compute the f1 score
This function computes the f1 score between
two vectors.
}

