% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{fnr}
\alias{fnr}
\alias{mtr_fnr}
\alias{mtr_miss_rate}
\alias{mtr_false_negative_rate}
\title{False Negative Rate}
\usage{
mtr_fnr(actual, predicted, cutoff = 0.5)

mtr_miss_rate(actual, predicted, cutoff = 0.5)

mtr_false_negative_rate(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}
}
\value{
A numeric output
}
\description{
False Negative Rate
}
\author{
An Chu
}
