% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{fpr}
\alias{fpr}
\alias{mtr_fpr}
\alias{mtr_fallout}
\alias{mtr_false_alarm_rate}
\alias{mtr_far}
\alias{mtr_false_positive_rate}
\title{False Positive Rate}
\usage{
mtr_fpr(actual, predicted, cutoff = 0.5)

mtr_fallout(actual, predicted, cutoff = 0.5)

mtr_false_alarm_rate(actual, predicted, cutoff = 0.5)

mtr_far(actual, predicted, cutoff = 0.5)

mtr_false_positive_rate(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}
}
\value{
A numeric scalar output
}
\description{
False Positive Rate
}
\author{
An Chu
}
