% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbif_utils.R
\name{get_abundance_type}
\alias{get_abundance_type}
\title{Classify species based on observation count}
\usage{
get_abundance_type(observation_count)
}
\arguments{
\item{observation_count}{Numeric. The number of observations for the species.}
}
\value{
A character string indicating the classification of the species.
}
\description{
This function classifies a species based on the number of observations
into categories like 'Extinct', 'Endangered', 'Rare', 'Common', or 'Overabundant'.
}
