% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invacost_utils.R
\name{get_total_cost_by_species}
\alias{get_total_cost_by_species}
\title{Get Total Cost by Species}
\usage{
get_total_cost_by_species(species_name, invacost_data)
}
\arguments{
\item{species_name}{A character string representing the species name to filter the data by.}

\item{invacost_data}{A data frame containing the cost data for multiple species, with at least two columns: \code{Species} and \code{Raw_cost_estimate_2017_USD_exchange_rate}.}
}
\value{
A numeric value representing the total cost for the specified species in billion USD. If no data is found for the species, returns 0.
}
\description{
This function calculates the total cost for a given species from a dataset containing raw cost estimates. The total cost is returned in billion USD.
}
\details{
The function filters the input data for the specified species and calculates the total raw cost estimate (in 2017 USD). The result is converted to billion USD and returned as a rounded value.
}
\examples{
# Example usage:
invacost_data <- data.frame(Species = c("species1", "species2", "species1"),
                            Raw_cost_estimate_2017_USD_exchange_rate = c(100, 200, 150))
get_total_cost_by_species("species1", invacost_data)

}
