% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{ll}
\alias{ll}
\title{Compute the log loss}
\usage{
ll(actual, predicted)
}
\arguments{
\item{actual}{binary ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
This function computes the elementwise log loss for a
number or a vector
}

