% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invacost_utils.R
\name{load_invacost_data}
\alias{load_invacost_data}
\title{Load InvaCost Dataset}
\usage{
load_invacost_data()
}
\value{
A data frame containing the InvaCost dataset.
}
\description{
This function loads the InvaCost dataset from the invacost package
and returns it to be assigned by a variable.
}
\examples{
invacost_data <- load_invacost_data()
}
