% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{mae}
\alias{mae}
\alias{mtr_mean_absolute_error}
\alias{mtr_mae}
\title{Mean Absolute Error}
\usage{
mtr_mean_absolute_error(actual, predicted)

mtr_mae(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} The ground truth numeric vector.}

\item{predicted}{\code{[numeric]} The predicted numeric vector, where each
element in the vector is a prediction of the corresponding elements in
\code{actual}.}
}
\value{
A numeric scalar output
}
\description{
\code{mtr_mean_absolute_error} is a measure of difference between the
predicted value and the true value.
}
\examples{

act <- c(3, -0.5, 2, 7)
pred <- c(2.5, 0.0, 2, 8)
mtr_mean_absolute_error(act, pred)
mtr_mae(act, pred)


}
\author{
An Chu
}
