\name{mapk}
\alias{mapk}
\title{Compute the mean average precision at k}
\usage{
  mapk(k, actual, predicted)
}
\arguments{
  \item{actual}{list of ground truth sets (vectors)}

  \item{predicted}{list of predicted sequences (vectors)}
}
\description{
  This function computes the mean average precision at k of
  two lists of sequences.
}

