% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\docType{package}
\name{metrics-package}
\alias{metrics}
\alias{metrics-package}
\title{metrics: Metrics For Evaluating Machine Learning Models}
\description{
Performance metrics for classification and regression problems
    that are widely used in supervised machine learning.
}
\author{
\strong{Maintainer}: An Chu \email{chuvanan.cva@gmail.com}

}
