% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{mpe}
\alias{mpe}
\title{MPE}
\usage{
mpe(x, target, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{target}{numeric vector}

\item{na.rm}{logical; should NAs be removed before calculation?}
}
\value{
An atomic numeric vector containing the calculated MPE
}
\description{
Calculates the Mean Percentage Error(MPE) between x and the
target.
}
\examples{
\dontrun{mpe(c(1, 2, 3))}            #> Error: different lengths (3, 0)
mpe(c(1, 2, 3), c(1, 4, 4))          #> 25.0
mpe(c(1, 2, NA, 4), c(2, 2, 5, NA))  #> NA
mpe(c(1, 2, NA, 4),
    c(2, 2, 5, NA),
    na.rm = TRUE)                    #> 12.5
}
