% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{mse}
\alias{mse}
\alias{mtr_mean_squared_error}
\alias{mtr_mse}
\title{Mean Squared Error}
\usage{
mtr_mean_squared_error(actual, predicted)

mtr_mse(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} The ground truth numeric vector.}

\item{predicted}{\code{[numeric]} The predicted numeric vector, where each
element in the vector is a prediction of the corresponding elements in
\code{actual}.}
}
\value{
A numeric scalar output
}
\description{
test
}
\author{
An Chu
}
