% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{msle}
\alias{msle}
\alias{mtr_mean_squared_log_error}
\alias{mtr_msle}
\title{Mean Squared Log Error}
\usage{
mtr_mean_squared_log_error(actual, predicted)

mtr_msle(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} The ground truth numeric vector.}

\item{predicted}{\code{[numeric]} The predicted numeric vector, where each
element in the vector is a prediction of the corresponding elements in
\code{actual}.}
}
\value{
A numeric scalar output
}
\description{
test
}
\examples{

act <- c(3, 5, 2.5, 7)
pred <- c(2.5, 5, 4, 8)
mtr_mean_squared_log_error(act, pred)


}
\author{
An Chu
}
