% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc.r
\name{mtr_auc_roc}
\alias{mtr_auc_roc}
\title{Area Under ROC}
\usage{
mtr_auc_roc(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}
}
\value{
A numeric scalar output
}
\description{
\code{mtr_auc_roc}
}
\examples{

set.seed(101)
pred <- runif(1000)
act <- round(pred)
pred[sample(1000, 500)] <- runif(500)   # noise
mtr_auc_prc(act, pred)


}
\seealso{
\code{\link{mtr_ar}} \code{\link{mtr_auc_prc}}
}
\author{
An Chu
}
