% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{mtr_average_precision}
\alias{mtr_average_precision}
\title{Average Precision}
\usage{
mtr_average_precision(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}
}
\value{
A numeric scalar output
}
\description{
Average Precision
}
\author{
An Chu
}
