% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier-score.r
\name{mtr_brier_score}
\alias{mtr_brier_score}
\title{Brier score}
\usage{
mtr_brier_score(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}
}
\value{
A scalar numeric output
}
\description{
Brier score measures the accuracy of probabilistic predictions; the smaller
the Brier score, the better. The Brier score always takes on a value between
zero and one.

Given N predictions, the Brier score is defined by the mean squared
difference between (1) the estimated probability assigned to the possible
outcome and (2) the actual outcome.

Here is a few examples that help the interpretation of Brier score.

Suppose that one is forecasting the probability P that it will rain on a
given day. Then the Brier score is calculated as follows:

\itemize{

\item If the forecast is 100\% (P = 1) and it rains, then the Brier Score
is 0, the best score achievable.

\item If the forecast is 100\% and it does not rain, then the Brier Score
is 1, the worst score achievable.

\item If the forecast is 70\% (P = 0.70) and it rains, then the Brier Score
is (0.70 − 1)^2 = 0.09.

\item If the forecast is 30\% (P = 0.30) and it rains, then the Brier Score
is (0.30 − 1)^2 = 0.49.

}
}
\examples{

act <- c(0, 1, 1, 0)
pred <- c(0.1, 0.9, 0.8, 0.3)
mtr_brier_score(act, pred)

}
\seealso{
\code{\link{mtr_log_loss}} \code{\link{mtr_mean_log_loss}}
}
\author{
An Chu
}
