% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{mtr_explained_variance}
\alias{mtr_explained_variance}
\title{Explained Variance}
\usage{
mtr_explained_variance(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} The ground truth numeric vector.}

\item{predicted}{\code{[numeric]} The predicted numeric vector, where each
element in the vector is a prediction of the corresponding elements in
\code{actual}.}
}
\value{
A numeric scalar output
}
\description{
\code{mtr_explained_variance} computes explained variance, also known as
explained variation, which is interpreted as the percentage of variation in
one numeric vector explained by another. The best possible score is 1.0,
lower values are worse.
}
\examples{

act <- c(3, -0.5, 2, 7)
pred <- c(2.5, 0.0, 2, 8)
mtr_explained_variance(act, pred)

act <- c(1.1, 1.9, 3.0, 4.4, 5.0, 5.6)
pred <- c(0.9, 1.8, 2.5, 4.5, 5.0, 6.2)
mtr_explained_variance(act, pred)

}
\seealso{
\code{\link{mtr_r2}}
}
\author{
An Chu
}
