% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{mtr_fbeta_score}
\alias{mtr_fbeta_score}
\title{F-beta Score}
\usage{
mtr_fbeta_score(actual, predicted, cutoff = 0.5, beta = 1)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}

\item{beta}{\code{[numeric]} To be filled}
}
\value{
A numeric scalar output
}
\description{
F-beta Score
}
\seealso{
\code{\link{mtr_f1score}}
}
\author{
An Chu
}
