% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{mtr_ks_statistic}
\alias{mtr_ks_statistic}
\title{KS Statistic}
\usage{
mtr_ks_statistic(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}
}
\value{
A numeric scalar output
}
\description{
Kolmogorov–Smirnov statistic quantifies a distance between the empirical
distribution function of the sample and the cumulative distribution function
of the reference distribution.
}
\author{
An Chu
}
