% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{mtr_r2}
\alias{mtr_r2}
\title{R2 Score}
\usage{
mtr_r2(actual, predicted)
}
\arguments{
\item{actual}{\code{[numeric]} The ground truth numeric vector.}

\item{predicted}{\code{[numeric]} The predicted numeric vector, where each
element in the vector is a prediction of the corresponding elements in
\code{actual}.}
}
\value{
A numeric scalar output
}
\description{
test
}
\examples{

act <- c(3, -0.5, 2, 7)
pred <- c(2.5, 0.0, 2, 8)
mtr_r2(act, pred)


}
\author{
An Chu
}
