% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invacost_utils.R
\name{normalize_costs}
\alias{normalize_costs}
\title{Normalize Cost Values}
\usage{
normalize_costs(df)
}
\arguments{
\item{df}{A data frame containing the InvaCost dataset.}
}
\value{
A data frame with the normalized cost values.
}
\description{
This function normalizes the cost values in the InvaCost dataset to a range of 0 to 1.
}
