% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{npv}
\alias{npv}
\alias{mtr_npv}
\alias{mtr_negative_predicted_value}
\title{Negative Predicted Value}
\usage{
mtr_npv(actual, predicted, cutoff = 0.5)

mtr_negative_predicted_value(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}
}
\value{
A numeric scalar output
}
\description{
Negative Predicted Value
}
\author{
An Chu
}
