% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{objective_threshold}
\alias{objective_threshold}
\title{Find the threshold(s) for a given objective and a given metric.}
\usage{
objective_threshold(data, metric, objective)
}
\arguments{
\item{data}{A dataframe with the metrics computed for different thresholds.}

\item{metric}{A string. The metric to optimize for. One of: "Accuracy",
"Precision", "Recall", "Specificity", "NPV" or "F1".}

\item{objective}{A numeric value. The objective to reach.}
}
\value{
A dataframe with the optimal threshold(s).
}
\description{
Find the threshold(s) for a given objective and a given metric.
}
\examples{
# add an example here
}
\seealso{
\code{\link{compute_predict_class}}, \code{\link{compute_metrics}},
  \code{\link{compute_optimal_threshold}}
}
