% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{qcut}
\alias{qcut}
\title{Compute the quantiles}
\usage{
qcut(data, pos_class = "Yes", n = 20L)
}
\arguments{
\item{data}{A dataframe with the probabilities for the positive class.}

\item{pos_class}{A string (default = "Yes"). The column's name of the
predicted probabilities.}

\item{n}{A numeric value (default = 20). The number of quantiles to compute.}
}
\value{
A dataframe with the quantiles computed.
}
\description{
Compute the quantiles.
}
\examples{
# add an example here
}
\seealso{
\code{\link{compute_lift}}, \code{\link{draw_lift}}
}
