% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{rae}
\alias{rae}
\title{Compute the relative absolute error}
\usage{
rae(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the relative absolute error between
two vectors
}

