% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{regression_params}
\alias{regression_params}
\title{Regression Metrics Parameters}
\arguments{
\item{actual}{\code{[numeric]} The ground truth numeric vector.}

\item{predicted}{\code{[numeric]} The predicted numeric vector, where each
element in the vector is a prediction of the corresponding elements in
\code{actual}.}
}
\description{
Documentation for shared parameters of functions that compute regression
metrics.
}
