\name{rmse}
\alias{rmse}
\title{Compute the root mean squared error#'
This function computes the root mean squared error
between two vectors}
\usage{
  rmse(actual, predicted)
}
\arguments{
  \item{actual}{ground truth vector}

  \item{predicted}{predicted vector}
}
\description{
  Compute the root mean squared error#' This function
  computes the root mean squared error between two vectors
}

