% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{rmsle}
\alias{rmsle}
\title{Compute the root mean squared log error}
\usage{
rmsle(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the root mean squared log error between
two vectors
}

