% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{rrse}
\alias{rrse}
\title{Compute the root relative squared error}
\usage{
rrse(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the root relative squared error between
two vectors
}

