\name{ScoreQuadraticWeightedKappa}
\alias{ScoreQuadraticWeightedKappa}
\title{Compute the quadratic weighted kappa}
\usage{
  ScoreQuadraticWeightedKappa(rater.a, rater.b, min.rating,
    max.rating)
}
\arguments{
  \item{rater.a}{is the first rater's ratings}

  \item{rater.b}{is the second rater's ratings}

  \item{min.rating}{is the minimum possible rating}

  \item{max.rating}{is the maximum possible rating}
}
\description{
  This function computes the quadratic weighted kappa
  between two vectors of integers
}

