% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{se}
\alias{se}
\title{Compute the squared error}
\usage{
se(actual, predicted)
}
\arguments{
\item{actual}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
This function computes the elementwise squared error for a
number or a vector
}

