% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{Shift}
\alias{Shift}
\title{Shift}
\usage{
Shift(x, n)
}
\arguments{
\item{x}{vector; the values to be shifted}

\item{n}{atomic integer; the number of places to Shift}
}
\value{
the shifted vector x (same length as the input vector)
}
\description{
Shifts the contents of a vector n places to the right (positive n) or left
(negative n).
}
\examples{
Shift(c(1, 2, 3), 1)  #> c(NA, 1, 2)
Shift(c(1, 2, 3), 0)  #> c(1, 2, 3)
Shift(c(1, 2, 3), -1) #> c(2, 3, NA)
}
