% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{smape}
\alias{smape}
\title{Symmetric Mean Absolute Percentage Error}
\usage{
smape(actual, predicted)
}
\arguments{
\item{actual}{ground truth numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
\code{smape} computes the symmetric mean absolute percentage error between
two numeric vectors
}
\details{
\code{smape} is defined as two times the average of \code{abs(actual - predicted) / (abs(actual) + abs(predicted))}.
As a result, it will provide \code{NaN} only if \code{actual} and \code{predicted}
are both zero. It has an upper bound of \code{2}, when either \code{actual} or
\code{predicted} are zero.
}
\examples{
actual <- c(1.1, 1.9, 3.0, 4.4, 5.0, 5.6)
predicted <- c(0.9, 1.8, 2.5, 4.5, 5.0, 6.2)
smape(actual, predicted)
}
\seealso{
\code{\link{mape}} \code{\link{mase}}
}
