% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{sse}
\alias{sse}
\title{Sum of Squared Errors}
\usage{
sse(actual, predicted)
}
\arguments{
\item{actual}{ground truth numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
\code{sse} computes the sum of squared errors between two numeric vectors
}
\examples{
actual <- c(1.1, 1.9, 3.0, 4.4, 5.0, 5.6)
predicted <- c(0.9, 1.8, 2.5, 4.5, 5.0, 6.2)
se(actual, predicted)
}
\seealso{
\code{\link{mse}}
}
