% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testing}
\alias{testing}
\title{testing}
\usage{
testing(x)
}
\arguments{
\item{x}{the input}
}
\value{
the output
}
\description{
This is a test function in my test package.
}
\examples{
testing("Anything!")
}
