% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{tnr}
\alias{tnr}
\alias{mtr_tnr}
\alias{mtr_true_negative_rate}
\alias{mtr_specificity}
\alias{mtr_selectivity}
\title{True Negative Rate}
\usage{
mtr_tnr(actual, predicted, cutoff = 0.5)

mtr_true_negative_rate(actual, predicted, cutoff = 0.5)

mtr_specificity(actual, predicted, cutoff = 0.5)

mtr_selectivity(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}
}
\value{
A numeric scalar output
}
\description{
True Negative Rate
}
\author{
An Chu
}
