% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-mat.r
\name{tpr}
\alias{tpr}
\alias{mtr_tpr}
\alias{mtr_true_positive_rate}
\alias{mtr_hit_rate}
\alias{mtr_sensitivity}
\alias{mtr_detection_rate}
\alias{mtr_recall}
\title{True Positive Rate}
\usage{
mtr_tpr(actual, predicted, cutoff = 0.5)

mtr_true_positive_rate(actual, predicted, cutoff = 0.5)

mtr_hit_rate(actual, predicted, cutoff = 0.5)

mtr_sensitivity(actual, predicted, cutoff = 0.5)

mtr_detection_rate(actual, predicted, cutoff = 0.5)

mtr_recall(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{\code{[numeric]} Ground truth binary numeric vector containing
1 for the positive class and 0 for the negative class.}

\item{predicted}{\code{[numeric]} A vector of estimated probabilities.}

\item{cutoff}{\code{[numeric]} A cutoff value for \code{predicted} vector
which classify a sample into a given class. Default value is 0.5}
}
\value{
A numeric scalar output
}
\description{
True Positive Rate
}
\author{
An Chu
}
