% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{ae}
\alias{ae}
\title{Compute the absolute error
This function computes the elementwise absolute error for a
number or a vector}
\usage{
ae(actual, predicted)
}
\arguments{
\item{actual}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
Compute the absolute error
This function computes the elementwise absolute error for a
number or a vector
}

