% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{ae}
\alias{ae}
\title{Absolute Error}
\usage{
ae(x, target, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector; the values to check}

\item{target}{numeric vector; the values to compare against}

\item{na.rm}{logical; should we remove NA cases?}
}
\value{
atomic numeric value; the absolute error
}
\description{
Calculates the absolute difference between \code{x} and \code{target}.
}
\examples{
ae(c(1, 2, 3), c(1, 2, 4))     #> 1
ae(c(1, 2, 3), c(2, 3, 4))     #> 3
ae(c(1, 2, 3), c(-1, -2, -3))  #> 12
}
