% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{ae}
\alias{ae}
\title{Compute the absolute error}
\usage{
ae(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the elementwise absolute error between
two numeric vectors
}
\examples{
actual <- c(1,1,1,0,0,0)
predicted <- c(1,0,1,1,0,0)
ae(actual, predicted)
}
