\name{apk}
\alias{apk}
\title{Compute the average precision at k}
\usage{
  apk(k, actual, predicted)
}
\arguments{
  \item{k}{max length of predicted sequence}

  \item{actual}{ground truth set (vector)}

  \item{predicted}{predicted sequence (vector)}
}
\description{
  This function computes the average precision at k between
  two sequences
}

