% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_retrieval.R
\name{apk}
\alias{apk}
\title{Average Precision at k}
\usage{
apk(k, actual, predicted)
}
\arguments{
\item{k}{The number of elements of \code{predicted} to consider in the calculation.}

\item{actual}{The ground truth vector of relevant documents. The vector can contain
any numeric or character values, order does not matter, and the
vector does not need to be the same length as \code{predicted}.}

\item{predicted}{The predicted vector of retrieved documents. The vector can
contain any numeric of character values. However, unlike \code{actual},
order does matter, with the most documents deemed most likely to
be relevant at the beginning.}
}
\description{
\code{apk} computes the average precision at k, in the context of information
retrieval problems.
}
\examples{
actual <- c('a', 'b', 'd')
predicted <- c('b', 'c', 'a', 'e', 'f')
apk(3, actual, predicted)
}
