% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{auc}
\alias{auc}
\title{Compute the area under the ROC (AUC)}
\usage{
auc(actual, predicted)
}
\arguments{
\item{actual}{binary vector}

\item{predicted}{real-valued vector that defines the ranking}
}
\description{
This function computes the area under the receiver-operator
characteristic (AUC)
}

