% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{auc}
\alias{auc}
\title{Compute the area under the ROC curve (AUC)}
\usage{
auc(actual, predicted)
}
\arguments{
\item{actual}{ground truth binary numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
This function computes the area under the receiver-operator
characteristic curve (AUC)
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(0.9, 0.8, 0.4, 0.5, 0.3, 0.2)
auc(actual, predicted)
}
