% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{ce}
\alias{ce}
\title{Classification Error}
\usage{
ce(actual, predicted)
}
\arguments{
\item{actual}{The ground truth vector, where elements of the vector can be any variable type.}

\item{predicted}{The predicted vector, where elements of the vector represent a
prediction for the corresponding value in \code{actual}.}
}
\description{
\code{ce} is defined as the proportion of elements in \code{actual} that are not equal
to the corresponding element in \code{predicted}.
}
\examples{
actual <- c('a', 'a', 'c', 'b', 'c')
predicted <- c('a', 'b', 'c', 'b', 'a')
ce(actual, predicted)
}
\seealso{
\code{\link{accuracy}}
}
