% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{ce}
\alias{ce}
\title{Compute the classification error}
\usage{
ce(actual, predicted)
}
\arguments{
\item{actual}{The ground truth vector.}

\item{predicted}{The predicted vector.}
}
\description{
This function computes the classification error
between two vectors
}
\examples{
actual <- c('a', 'a', 'c', 'b', 'c')
predicted <- c('a', 'b', 'c', 'b', 'a')
ce(actual, predicted)
}
