% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute several metrics for classification tasks.}
\usage{
compute_metrics(data, threshold = 0.5, obs = "obs", prob = "prob",
  pos_class = "Yes", neg_class = "No")
}
\arguments{
\item{data}{A dataframe with the observed and the predicted class.}

\item{threshold}{A numerical value or a sequence of numerical values (default
= 0.5). The threshold(s) to compute the metrics.}

\item{obs}{A string (default = "obs"). The column's name of the observed
class.}

\item{prob}{A string (default = "prob"). The column's name of the predicted
probabilities.}

\item{pos_class}{A string (default = "Yes"). How the positive class is coded.}

\item{neg_class}{A string (default = "No"). How the negative class is coded.}
}
\value{
A dataframe with the computed metrics.
}
\description{
Compute the following metrics: "Accuracy", "Precision", "Recall",
"Specificity", "NPV" and "F1".
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  prob = c(runif(n = 20, min = 0.3, max = 0.8),
           runif(n = 20, min = 0.1, max = 0.6))
)

compute_metrics(data, threshold = 0.5)
compute_metrics(data, threshold = seq(0, 1, 0.1))
}
\seealso{
\code{\link{compute_predict_class}},
  \code{\link{compute_optimal_threshold}}, \code{\link{objective_threshold}}
}
