% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_optimal_threshold}
\alias{compute_optimal_threshold}
\title{Compute the optimal threshold for the accuracy or the F1-score}
\usage{
compute_optimal_threshold(data, metric = c("Accuracy", "F1"))
}
\arguments{
\item{data}{A dataframe with the metrics computed for different thresholds.}

\item{metric}{A string. The metric to maximize. One of "Accuracy" or "F1".}
}
\value{
A dataframe with the optimal threshold(s).
}
\description{
Compute the optimal threshold for the accuracy or the F1-score.
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  Yes = c(runif(n = 20, min = 0.3, max = 0.8),
          runif(n = 20, min = 0.1, max = 0.6))
)

predict_class <- lapply(c(0.3, 0.5, 0.7), function(threshold) {
  compute_predict_class(data, threshold = threshold)
})

metrics <- data.table::setDF(
  data.table::rbindlist(
    lapply(predict_class, compute_metrics)
  )
)
metrics[["threshold"]] <- c(0.3, 0.5, 0.7)

compute_optimal_threshold(metrics, metric = "F1")
}
\seealso{
\code{\link{compute_predict_class}}, \code{\link{compute_metrics}},
  \code{\link{objective_threshold}}
}
