% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{draw_metrics}
\alias{draw_metrics}
\title{Plot the selected metrics}
\usage{
draw_metrics(data, metric = c("Accuracy", "Precision", "Recall",
  "Specificity", "NPV", "F1"))
}
\arguments{
\item{data}{A dataframe or a list of dataframes with the metrics computed.}

\item{metric}{A string or a vector of strings. The metric(s) to plot.}
}
\value{
A ggplot2 object.
}
\description{
Plot the selected metrics.
}
\examples{
# add an example here
}
\seealso{
\code{\link{compute_predict_class}}, \code{\link{compute_metrics}}
}
