% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{draw_metrics}
\alias{draw_metrics}
\title{Plot the selected metrics}
\usage{
draw_metrics(data, metric = c("Accuracy", "Precision", "Recall",
  "Specificity", "NPV", "F1"), threshold = "threshold")
}
\arguments{
\item{data}{A dataframe or a list of dataframes with the metrics computed.}

\item{metric}{A string or a vector of string. The metric(s) to plot.}

\item{threshold}{A string (default = "pred"). The column's name of the
threshold.}
}
\value{
A ggplot2 object.
}
\description{
Plot the selected metrics.
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  Yes = c(runif(n = 20, min = 0.3, max = 0.8),
          runif(n = 20, min = 0.1, max = 0.6))
)

predict_class <- lapply(c(0.3, 0.5, 0.7), function(threshold) {
  compute_predict_class(data, threshold = threshold)
})

metrics <- data.table::setDF(
  data.table::rbindlist(
    lapply(predict_class, compute_metrics)
  )
)
metrics[["threshold"]] <- c(0.3, 0.5, 0.7)

draw_metrics(metrics)
}
\seealso{
\code{\link{compute_predict_class}}, \code{\link{compute_metrics}}
}
