% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{draw_roc}
\alias{draw_roc}
\title{Plot the ROC curve}
\usage{
draw_roc(data, recall = "Recall", specificity = "Specificity")
}
\arguments{
\item{data}{A dataframe or a list of dataframes with the recall and the
specificity.}

\item{recall}{A string (default = "Recall"). The column's name of the
computed recall.}

\item{specificity}{A string (default = "Specificity"). The column's name of
the computed specificity.}
}
\value{
A ggplot2 object.
}
\description{
Plot the ROC curve.
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  prob = c(runif(n = 20, min = 0.3, max = 0.8),
           runif(n = 20, min = 0.1, max = 0.6))
)

metrics <- compute_metrics(data, threshold = c(0.3, 0.7, 0.2))

draw_roc(metrics)
}
\seealso{
\code{\link{compute_predict_class}}, \code{\link{compute_metrics}}
}
