\name{f1}
\alias{f1}
\title{Compute the f1 score}
\usage{
  f1(actual, predicted)
}
\arguments{
  \item{actual}{ground truth vector}
  
  \item{predicted}{predicted vector}
}
\description{
  This function computes the f1 score given the actual and predicted vectors.
  f1 score is calculated after calculating Precision and Recall from the two vectors. 
}
