% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{f1}
\alias{f1}
\title{Compute the f1 score}
\usage{
f1(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the f1 score between
two numeric vectors
}
\examples{
actual <- c(1,1,1,0,0,0)
predicted <- c(1,0,1,1,0,0)
f1(actual, predicted)
}
