% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{ll}
\alias{ll}
\title{Compute the log loss}
\usage{
ll(actual, predicted)
}
\arguments{
\item{actual}{ground truth binary numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
This function computes the elementwise log loss between
two numeric vectors
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(0.9, 0.8, 0.4, 0.5, 0.3, 0.2)
ll(actual, predicted)
}
