% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{logLoss}
\alias{logLoss}
\title{Compute the mean log loss}
\usage{
logLoss(actual, predicted)
}
\arguments{
\item{actual}{binary ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the mean log loss between
two vectors
}

