% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{logLoss}
\alias{logLoss}
\title{Mean Log Loss}
\usage{
logLoss(actual, predicted)
}
\arguments{
\item{actual}{The ground truth binary numeric vector containing 1 for the positive
class and 0 for the negative class.}

\item{predicted}{A numeric vector of predicted values, where the values correspond
to the probabilities that each observation in \code{actual}
belongs to the positive class}
}
\description{
\code{logLoss} computes the average log loss between two numeric vectors.
}
\examples{
actual <- c(1, 1, 1, 0, 0, 0)
predicted <- c(0.9, 0.8, 0.4, 0.5, 0.3, 0.2)
logLoss(actual, predicted)
}
\seealso{
\code{\link{ll}}
}
