\name{mae}
\alias{mae}
\title{Compute the mean absolute error#'
This function computes the mean absolte error between
two vectors}
\usage{
  mae(actual, predicted)
}
\arguments{
  \item{actual}{ground truth vector}

  \item{predicted}{vector}
}
\description{
  Compute the mean absolute error#' This function computes
  the mean absolte error between two vectors
}

