% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{mae}
\alias{mae}
\title{Compute the mean absolute error}
\usage{
mae(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the mean absolute error between
two numeric vectors
}
\examples{
actual <- c(1,1,1,0,0,0)
predicted <- c(1,0,1,1,0,0)
mae(actual, predicted)
}
