% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{mape}
\alias{mape}
\title{MAPE}
\usage{
mape(x, target, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector; the forecast to evaluate}

\item{target}{numeric vector; the target values}

\item{na.rm}{logical; should NAs be removed before calculation?}
}
\value{
An atomic numeric vector containing the calculated MAPE
}
\description{
Calculates the Mean Absolute Percentage Error(MAPE) between x and the
target.
}
\examples{
\dontrun{mape(c(1, 2, 3))}            #> Error: different lengths (3, 0)
mape(c(1, 2, 3), c(1, 4, 4))          #> 25.0
mape(c(1, 2, NA, 4), c(2, 2, 5, NA))  #> NA
mape(c(1, 2, NA, 4),
     c(2, 2, 5, NA),
     na.rm = TRUE)                    #> 12.5
}
