% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{mase}
\alias{mase}
\title{Compute the mean absolute scaled error}
\usage{
mase(actual, predicted, step_size = 1)
}
\arguments{
\item{actual}{ground truth numeric vector ordered in time, with most recent
observation at the end of the vector}

\item{predicted}{predicted numeric vector ordered in time.}

\item{step_size}{an integer that specifies how to many observations to look back
in order to compute the naive forecast.}
}
\description{
This function computes the mean absolute scaled error between two numeric
vectors. This function is only intended for time series data, where
\code{actual} and \code{numeric} are numeric vectors ordered by time.
}
