\name{MeanQuadraticWeightedKappa}
\alias{MeanQuadraticWeightedKappa}
\title{Compute the mean quadratic weighted kappa}
\usage{
  MeanQuadraticWeightedKappa(kappas, weights)
}
\arguments{
  \item{kappas}{is a vector of possible kappas}

  \item{weights}{is an optional vector of ratings}
}
\description{
  This function computes the mean quadratic weighted kappa,
  which can optionally be weighted
}

