% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{MeanQuadraticWeightedKappa}
\alias{MeanQuadraticWeightedKappa}
\title{Compute the mean quadratic weighted kappa}
\usage{
MeanQuadraticWeightedKappa(kappas, weights)
}
\arguments{
\item{kappas}{is a vector of possible kappas}

\item{weights}{is an optional vector of ratings}
}
\description{
This function computes the mean quadratic weighted
kappa, which can optionally be weighted
}
\examples{
kappas <- c(0.3,0.2,0.2,0.5,0.1,0.2)
weights <- c(1.0,2.5,1.0,1.0,2.0,3.0)
MeanQuadraticWeightedKappa(kappas, weights)
}
