% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{MeanQuadraticWeightedKappa}
\alias{MeanQuadraticWeightedKappa}
\title{Mean Quadratic Weighted Kappa}
\usage{
MeanQuadraticWeightedKappa(kappas, weights = rep(1, length(kappas)))
}
\arguments{
\item{kappas}{A numeric vector of possible kappas.}

\item{weights}{An optional numeric vector of ratings.}
}
\description{
\code{MeanQuadraticWeightedKappa} computes the mean quadratic weighted kappa,
which can optionally be weighted
}
\examples{
kappas <- c(0.3 ,0.2, 0.2, 0.5, 0.1, 0.2)
weights <- c(1.0, 2.5, 1.0, 1.0, 2.0, 3.0)
MeanQuadraticWeightedKappa(kappas, weights)
}
\seealso{
\code{\link{ScoreQuadraticWeightedKappa}}
}
