\name{mse}
\alias{mse}
\title{Compute the mean squared error#'
This function computes the mean squared error between
two vectors}
\usage{
  mse(actual, predicted)
}
\arguments{
  \item{actual}{ground truth vector}

  \item{predicted}{predicted vector}
}
\description{
  Compute the mean squared error#' This function computes
  the mean squared error between two vectors
}

