% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{mse}
\alias{mse}
\title{MSE}
\usage{
mse(x, target, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{target}{numeric vector}

\item{na.rm}{logical; should NAs be removed before calculation?}
}
\value{
An atomic numeric vector containing the calculated MSE
}
\description{
Calculates the Mean Squared Error(MSE), or Mean Squared Deviation (MSD),
between x and the target. Calculation uses (target - x).
}
\examples{
mse(c(1, 2, 3), c(1, 4, 4))      #> 1.666667
mse(c(1, 2, NA), c(NA, 10, 20))  #> NA
}
