% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{mse}
\alias{mse}
\title{Compute the mean squared error}
\usage{
mse(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the mean squared error between
two numeric vectors
}
\examples{
actual <- c(1,1,1,0,0,0)
predicted <- c(0.9,0.8,0.4,0.5,0.3,0.2)
mse(actual, predicted)
}
