% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{msle}
\alias{msle}
\title{Compute the mean squared log error}
\usage{
msle(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the mean squared log error between
two vectors
}

