% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{msle}
\alias{msle}
\title{Mean squared log error}
\usage{
msle(actual, predicted)
}
\arguments{
\item{actual}{ground truth non-negative vector}

\item{predicted}{predicted non-negative vector}
}
\description{
\code{msle} computes the mean squared log error between two numeric vectors
}
\details{
\code{msle} adds one to both \code{actual} and \code{predicted} before taking
the natural logarithm to avoid taking the natural log of zero.
}
\examples{
actual <- c(1.1, 1.9, 3.0, 4.4, 5.0, 5.6)
predicted <- c(0.9, 1.8, 2.5, 4.5, 5.0, 6.2)
msle(actual, predicted)
}
\seealso{
\code{\link{rmsle}}
}
