% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{objective_threshold}
\alias{objective_threshold}
\title{Find the threshold(s) for a given objective and a given metric.}
\usage{
objective_threshold(data, metric, objective)
}
\arguments{
\item{data}{A dataframe with the metrics computed for different thresholds.}

\item{metric}{A string. The metric to consider. One of: "Accuracy",
"Precision", "Recall", "Specificity", "NPV" or "F1".}

\item{objective}{A numeric value. The objective to reach.}
}
\value{
A dataframe with the optimal threshold(s).
}
\description{
Find the threshold(s) for a given objective and a given metric.
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  prob = c(runif(n = 20, min = 0.3, max = 0.8),
           runif(n = 20, min = 0.1, max = 0.6))
)

metrics <- compute_metrics(data, threshold = c(0.3, 0.7, 0.2))

objective_threshold(metrics, metric = "Precision", objective = 0.75)
}
\seealso{
\code{\link{compute_predict_class}}, \code{\link{compute_metrics}},
  \code{\link{compute_optimal_threshold}}
}
