% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_classification.R
\name{params_binary}
\alias{params_binary}
\title{Inherit Documentation for Binary Classification Metrics}
\arguments{
\item{actual}{ground truth binary numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
This object provides the documentation for the parameters of functions
             that provide binary classification metrics
}
