% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\docType{data}
\name{params_classification}
\alias{params_classification}
\title{Inherit Documentation for Classification Metrics}
\format{An object of class \code{NULL} of length 0.}
\usage{
params_classification
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This object provides the documentation for the parameters of functions
             that provide classification metrics
}
\keyword{datasets}
