% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{params_classification}
\alias{params_classification}
\title{Inherit Documentation for Classification Metrics}
\arguments{
\item{actual}{The ground truth vector.}

\item{predicted}{The predicted vector.}
}
\description{
This object provides the documentation for the parameters of functions
             that provide classification metrics
}
