% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{params_classification}
\alias{params_classification}
\title{Inherit Documentation for Classification Metrics}
\arguments{
\item{actual}{The ground truth vector, where elements of the vector can be any variable type.}

\item{predicted}{The predicted vector, where elements of the vector represent a
prediction for the corresponding value in \code{actual}.}
}
\description{
This object provides the documentation for the parameters of functions
             that provide classification metrics
}
