% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{params_regression}
\alias{params_regression}
\title{Inherit documentation for regression metrics}
\arguments{
\item{actual}{ground truth numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
This object provides the documentation for the parameters of functions
             that provide regression metrics
}
