% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\docType{data}
\name{params_regression}
\alias{params_regression}
\title{Inherit documentation for regression metrics}
\format{An object of class \code{NULL} of length 0.}
\usage{
params_regression
}
\arguments{
\item{actual}{ground truth numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
This object provides the documentation for the parameters of functions
             that provide regression metrics
}
\keyword{datasets}
