% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{params_regression}
\alias{params_regression}
\title{Inherit Documentation for Regression Metrics}
\arguments{
\item{actual}{The ground truth numeric vector.}

\item{predicted}{The predicted numeric vector, where each element in the vector
is a prediction for the corresponding element in \code{actual}.}
}
\description{
This object provides the documentation for the parameters of functions
             that provide regression metrics
}
