% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{qcut}
\alias{qcut}
\title{Compute the quantiles}
\usage{
qcut(data, prob = "prob", nb_quantiles = 20L)
}
\arguments{
\item{data}{A dataframe with the probabilities for the positive class.}

\item{prob}{A string (default = "prob"). The column's name of the predicted
probabilities.}

\item{nb_quantiles}{An integer (default = 20). The number of quantiles to
compute.}
}
\value{
A dataframe with the quantiles computed.
}
\description{
Compute the quantiles.
}
\examples{
data <- data.frame(
  obs = c(rep("Yes", 20), rep("No", 20)),
  prob = c(runif(n = 20, min = 0.3, max = 0.8),
           runif(n = 20, min = 0.1, max = 0.6))
)

quantiles <- qcut(data, prob = "prob")
}
\seealso{
\code{\link{compute_lift}}, \code{\link{draw_lift}}
}
