% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{rae}
\alias{rae}
\title{Compute the relative absolute error}
\usage{
rae(actual, predicted)
}
\arguments{
\item{actual}{ground truth numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
This function computes the relative absolute error between
two numeric vectors
}
\examples{
actual <- c(1,1,1,0,0,0)
predicted <- c(1,0,1,1,0,0)
rae(actual, predicted)
}
