% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(x, target, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{target}{numeric vector}

\item{na.rm}{logical; should NAs be removed before calculation?}
}
\value{
An atomic numeric vector containing the calculated RMSE
}
\description{
Calculates the Root Mean Squared Error(RMSE) between x and the target.
Calculation uses (target - x).
}
\examples{
rmse(c(1, 2, 3), c(1, 4, 4))      #> 1.290994
rmse(c(1, 2, NA), c(NA, 10, 20))  #> NA
}
