% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{rrse}
\alias{rrse}
\title{Compute the root relative squared error}
\usage{
rrse(actual, predicted)
}
\arguments{
\item{actual}{ground truth numeric vector}

\item{predicted}{predicted numeric vector}
}
\description{
This function computes the root relative squared error between
two numeric vectors
}
\examples{
actual <- c(1,1,1,0,0,0)
predicted <- c(0.9,0.8,0.4,0.5,0.3,0.2)
rrse(actual, predicted)
}
