% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{rse}
\alias{rse}
\title{Compute the relative squared error}
\usage{
rse(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the relative squared error between
two vectors
}

