% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{ScoreQuadraticWeightedKappa}
\alias{ScoreQuadraticWeightedKappa}
\title{Compute the quadratic weighted kappa}
\usage{
ScoreQuadraticWeightedKappa(rater.a, rater.b, min.rating, max.rating)
}
\arguments{
\item{rater.a}{is the first rater's ratings}

\item{rater.b}{is the second rater's ratings}

\item{min.rating}{is the minimum possible rating}

\item{max.rating}{is the maximum possible rating}
}
\description{
This function computes the quadratic weighted kappa
between two vectors of integers
}
\examples{
rater.a <- c(1,4,5,5,2,1)
rater.b <- c(2,2,4,5,3,3)
ScoreQuadraticWeightedKappa(rater.a, rater.b, 1, 5)
}
