% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{se}
\alias{se}
\title{Compute the squared error}
\usage{
se(actual, predicted)
}
\arguments{
\item{actual}{ground truth vector}

\item{predicted}{predicted vector}
}
\description{
This function computes the elementwise squared error between
two numeric vectors
}
\examples{
actual <- c(1,1,1,0,0,0)
predicted <- c(0.9,0.8,0.4,0.5,0.3,0.2)
se(actual, predicted)
}
