% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{Shift}
\alias{Shift}
\title{Shift (Internal)}
\usage{
Shift(x, n)
}
\arguments{
\item{x}{vector; the values to be shifted}

\item{n}{atomic integer; the number of places to Shift}
}
\value{
the shifted vector \code{x} (same length as the input vector)
}
\description{
Shifts the contents of a vector \code{n} places to the right (positive
\code{n}) or left (negative \code{n}).
}
