% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{sle}
\alias{sle}
\title{Compute the squared log error}
\usage{
sle(actual, predicted)
}
\arguments{
\item{actual}{ground truth number or vector}

\item{predicted}{predicted number or vector}
}
\description{
This function computes the elementwise squared log error for a
number or a vector
}

