% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{sle}
\alias{sle}
\title{Squared Log Error}
\usage{
sle(actual, predicted)
}
\arguments{
\item{actual}{ground truth non-negative vector}

\item{predicted}{predicted non-negative vector}
}
\description{
\code{sle} computes the elementwise squared log error between two numeric vectors
}
\details{
\code{sle} adds one to both \code{actual} and \code{predicted} before taking
the natural logarithm to avoid taking the natural log of zero.
}
\examples{
actual <- c(1.1, 1.9, 3.0, 4.4, 5.0, 5.6)
predicted <- c(0.9, 1.8, 2.5, 4.5, 5.0, 6.2)
sle(actual, predicted)
}
\seealso{
\code{\link{msle}} \code{\link{rmsle}}
}
