% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{sse}
\alias{sse}
\title{SSE}
\usage{
sse(x, target, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{target}{numeric vector}

\item{na.rm}{logical; should NAs be removed before the calculation?}
}
\value{
An atomic numeric vector containing the calculated SSE
}
\description{
Calculates the sum of squared errors (SSE) between x and the target.
}
\examples{
sse(c(1, 2, 3), c(2, 3, 4))                 #> 1 + 1 + 1 =  3.0
sse(c(1, 2, 3), c(3, 4, 5))                 #> 4 + 4 + 4 = 12.0
sse(c(1, 2, NA), c(1, 2, 3))                #> NA
sse(c(1, 2, NA), c(1, 2, 3), na.rm = TRUE)  #> 0 + 0 = 0.0
}
